refactor: package 생성 및 테스트 코드 분리
